Directly integrate parts of libproc to this codebase
The 4.x release of libprocps, aka libproc2, no longer exposes API
calls that are needed by idleinject. The rationale from libprocps
developer was that those never should have been exposed and were
only intended for internal use by the library.

It was more straightforward to integrate the necessary sources to
this codebase than to modify the idleinject application to work with
the new library release. This was done as opposed to requiring the
older 3.x library version at runtime to make building this more
portable.

Because this now integrates LGPL code directly, the project license
from this point forward is now LGPL2.1 to match libprocps' library.

Signed-off-by: Kris Bahnsen <kris@embeddedTS.com>